LIBITD-1447. Refactored total_row_counts and verify_empty_tables
Refactored "total_row_counts" and "verify_empty_tables" functionality
into top-level script.s

https://issues.umd.edu/browse/LIBITD-1447
