+++ redcapdatapurge/__main__.py
@@ -1,5 +1,3 @@
-import dataset
-
 import os
 from dotenv import load_dotenv
 load_dotenv()
@@ -178,39 +176,6 @@ def cleanup_sql_for_other_orphans():
     return sql_statements
 
 
-def verify_empty_tables(table_row_count):
-    """
-    Verifies that the expected tables have no entries, as they are not
-    processed by the script.
-    """
-    expected_empty_tables = {
-        'redcap_external_modules_log',
-        'redcap_external_modules_log_parameters',
-        'redcap_instrument_zip',
-        'redcap_instrument_zip_authors',
-        'redcap_ip_banned',
-        'redcap_projects_external',
-        'redcap_pub_articles',
-        'redcap_pub_authors',
-        'redcap_pub_matches',
-        'redcap_pub_mesh_terms',
-        'redcap_user_whitelist'
-    }
-
-    for t in expected_empty_tables:
-        try:
-            count_for_table = table_row_count[t]
-            if count_for_table is None:
-                print(f"ERROR: Expected empty table '{t}' not found in row count.")
-                exit(1)
-            if count_for_table != 0:
-                print(f"Error: Expected empty table '{t}' had {count_for_table} rows.")
-                exit(1)
-        except KeyError:
-            print(f"ERROR: Expected empty table '{t}' not found in row count.")
-            exit(1)
-
-
 def purge_sql_for_redcap_projects(project_ids_to_keep):
     sql_statements = []
 
@@ -280,8 +245,9 @@ def purge_sql_unattached_tables_with_user_name(user_names_to_keep):
 
 
 def main():
-    required_env_vars = ["DB_URL", "SUMMARY_OUTPUT_FILE",
-                         "PURGE_QUERIES_OUTPUT_FILE", "RELATED_ITEM_IDS_OUTPUT_FILE"]
+    required_env_vars = ["DB_URL", "CLEANUP_OUTPUT_FILE",
+                         "PURGE_QUERIES_OUTPUT_FILE",
+                         "REDCAP_ADMIN_PURGE_QUERIES_OUTPUT_FILE"]
     for var in required_env_vars:
         if not os.getenv(var):
             print(f"Required environment variable '{var}' is missing.")
@@ -296,14 +262,6 @@ def main():
     cleanup_file = os.getenv("CLEANUP_OUTPUT_FILE")
     with open(cleanup_file, 'w') as fp:
         fp.write('\n'.join(cleanup_sql_statements))
-
-    # Get values from .env config file
-    db_url = os.getenv("DB_URL")
-    db = dataset.connect(db_url)
-
-    total_rows_count = calculate_total_rows_count(db)
-    print(f"total_rows_count: {total_rows_count}")
-    verify_empty_tables(total_rows_count)
 
     usernames_to_keep = [
         'dyarnell', 'dmilton', 'sbobbin', 'jbueno', 'balbert', 'rwashing',
@@ -333,6 +291,8 @@ def main():
     purge_queries_file = os.getenv("PURGE_QUERIES_OUTPUT_FILE")
     with open(purge_queries_file, 'w') as fp:
         fp.write('\n'.join(purge_sql_statements))
+
+
 
 
 if __name__ == '__main__':
